+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## 0.1.2 (TBD)
 
+### Changes
+
+* [#108][gh-108] Locally overwrite the link following behaviour in the org-roam-buffer to open files in the same window `org-roam` was called from
+
 ### Breaking Changes
 * [#103][gh-103] Change `org-roam-file-format` to a function: `org-roam-file-name-function` to allow more flexible file name customizaton. Also changes `org-roam-use-timestamp-as-filename` to `org-roam-filename-noconfirm` to better describe what it does.
 
@@ -53,6 +57,7 @@ Mostly a documentation/cleanup release.
 [gh-98]: https://github.com/jethrokuan/org-roam/pull/98
 [gh-103]: https://github.com/jethrokuan/org-roam/pull/103
 [gh-105]: https://github.com/jethrokuan/org-roam/pull/105
+[gh-108]: https://github.com/jethrokuan/org-roam/pull/108
 
  # Local Variables:
  # eval: (auto-fill-mode -1)